Cloud functions code/Compose Image/requirements.txt
azure-storage-blob==12.8.1
azure-cognitiveservices-vision-computervision==0.7.0
azure-cosmos==4.2.0
openai==0.7.0
Pillow==8.2.0
flask==2.0.1
requests==2.25.1
msrest==0.7.1
azure-identity